Make facebook SDK to work with turbolinks
Turbolinks doesn't reload page on each page visit. It just changes only
body of document. So it doesn't reexecute sources listed in head.

When facebook SDK is added to head, it changes #fb-root element in
body. In order to make Facebook SDK work on multiple pages w/o
reexecuting script in head we have to keep #fb-root consistent on all
the pages.